requirements.txt
dspy==2.5.43
chromadb
openai
python-dotenv
pandas 
sqlalchemy 
streamlit
gradio
ollama

# hugging face
transformers